simplify version specific files
